internal/schema: Replace TupleConsExpr with SetExpr
The attributes in question will never be treated as data, so the data type doesn't really matter as much in this context. The main effect is hints in the completion and hover data.

Set seems most appropriate of all options (list, set, tuple), given that every element is of the same type (which rules out tuple) and order of elements doesn't matter (which rules out list).