Conclude Terraform Experiments Completion
This commit concludes the work on the Terraform experiments feature by removing the list of features from completion for the `experiments` attriubute.

Terraform language experiments are temporary features that are added, removed, and/or concluded at cadences too unpredictable to be included in this completion list. These are used in alpha or beta builds of Terraform and are changed frequently enough that it would either always be out of date or require constant releases to keep up to date.

As such, the completion list for the `experiments` attribute has been removed and documentation added to instruct the user to look to documentation for the particular release they are using for a list of features that can be enabled.