Remove InstalledModuleCalls merging logic
As we are able to build the schema just by looking at the declared
module blocks, we can stop iterating over the installed modules. Since
the LS rearchitecture we stopped decode the whole .terraform folder, so
there isn't much to add here anyway. This should give us a small memory
and performance improvement.