(TF-18286) Add support for Stack metadata (#364)
This adds support for loading stack metadata from a set of Stack files. The metadata is loaded into a `stack.Meta` struct which contains the path to the stack, the filenames of the files that were loaded, and maps of components, variables, and outputs. The `stack.Meta` struct also contains a map of provider requirements which is a map of provider addresses to version constraints.

---------

Co-authored-by: Ansgar Mertens <ansgar@hashicorp.com>
Co-authored-by: Daniel Banck <dbanck@users.noreply.github.com>