Raise HCL diagnostics
This commit raises HCL diagnostics discovered during early decoding of the stack files. Previously the diagnostics were appended to a slice of diagnostics, which did not track which file it refered to. Now they are stored in a map of diagnostics, where the key is the filename of the stack file.